refactor: extract utilities from Home component
